requirements.txt
Django==4.2.8
psycopg2-binary==2.9.9
python-decouple==3.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
pyzk==0.9
django-celery-beat==2.5.0
django-filter==23.5
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3
